+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 Changelog
 ==============
 
@@ -8,6 +9,7 @@ The releases of major and minor versions contain an overview of changes since th
 Version 1.6.x
 -------------
 ## Version 1.6.4 (20xx/xx)
+- Added support for PRISM models that use unbounded integer variables.
 - Added an export of check results to json. Use `--exportresult` in the command line interface.
 - Added computation of steady state probabilities for DTMC/CTMC in the sparse engine. Use `--steadystate` in the command line interface.
 - Implemented parsing and model building of Stochastic multiplayer games (SMGs) in the PRISM language. No model checking implemented, for now.